# AstraServerGo.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AstraServerGo",
    "description": "基于 `AstraScheduleServerGo` 源码生成的 OpenAPI 3.0 文档。\n\n- 共 55 个 API 端点，覆盖客户端课表、天气、WebSocket、管理后台认证、用户管理、学校/年级/班级管理、配置管理、自动任务、倒数日、备份导入导出等功能。\n- JWT 认证接口使用 `jwtAuth` 安全方案（通过 POST /web/auth/login 获取 Token）。\n- 写入接口同时支持 BasicAuth（兼容旧版客户端）和 JWT + 密码验证。\n- 部分接口返回结构是动态 JSON（`map[string]interface{}`），已尽量按实际行为建模。\n",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/web/menu": {
      "get": {
        "summary": "获取管理端菜单",
        "operationId": "getMenu",
        "tags": ["web-menu"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "菜单树",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MenuNode" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": []
      }
    },
    "/web/structure": {
      "get": {
        "summary": "获取学校-年级-班级结构",
        "operationId": "getStructure",
        "tags": ["web-menu"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "结构树",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/StructureNode" }
                }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/web/autorun": {
      "get": {
        "summary": "获取自动任务列表",
        "description": "返回所有自动任务记录，含过期/生效中/待生效状态判定。",
        "operationId": "getAutorunStatus",
        "tags": ["web-autorun"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "自动任务列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AutorunRecordView" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/web/countdown": {
      "get": {
        "summary": "获取倒数日配置列表",
        "description": "可通过 `scope` 查询参数过滤。",
        "operationId": "getCountdownStatus",
        "tags": ["web-countdown"],
        "parameters": [
          {
            "name": "scope", "in": "query", "description": "作用域过滤，格式 school/grade/class",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "倒数日配置列表",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CountdownListResponse" }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "新增/更新倒数日配置",
        "description": "当 `id` 为空时服务端会根据 scope+schedules 生成稳定 ID。需要 BasicAuth 认证。",
        "operationId": "putCountdownRule",
        "tags": ["web-countdown"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CountdownPayload" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "写入成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AutorunPutSuccess" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
    "/web/users": {
      "get": {
        "summary": "获取用户列表",
        "operationId": "listUsers",
        "tags": ["web-users"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/UserInfoFull" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      },
      "post": {
        "summary": "创建用户",
        "operationId": "createUser",
        "tags": ["web-users"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password", "role"],
                "properties": {
                  "username": { "type": "string", "minLength": 3 },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "school_w", "grade_w", "class_w", "readonly"] },
                  "scope": { "type": "string", "description": "作用域，格式 school/grade/class" },
                  "must_change_pwd": { "type": "boolean", "description": "是否强制修改密码" },
                  "must_change_username": { "type": "boolean", "description": "是否强制修改用户名" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "user"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "user": { "$ref": "#/components/schemas/UserInfo" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "未认证或无 admin 权限" },
          "409": {
            "description": "用户名已存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MenuNode": {
        "type": "object",
        "properties": {
          "to": { "type": "string", "description": "路由路径" },
          "text": { "type": "string", "description": "显示文本" },
          "key": { "type": "string" },
          "raw": { "type": "string" },
          "children": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MenuNode" },
            "nullable": true
          }
        }
      },
      "StructureNode": {
        "type": "object",
        "properties": {
          "text": { "type": "string" },
          "children": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StructureNode" },
            "nullable": true
          }
        }
      },
...
      "ScopeField": {
        "description": "作用域字段，支持单个 scope 字符串（school/grade/class）或字符串数组，也可为空表示 ALL。",
        "oneOf": [
          { "type": "string" },
          { "type": "array", "items": { "type": "string" } }
        ]
      },
      "AutorunRecordView": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "HashID" },
          "type": {
            "type": "string",
            "description": "任务类型：COMPENSATION(0) | TIMETABLE(1) | SCHEDULE(2) | ALL(3)"
          },
          "priority": { "type": "integer" },
          "status": {
            "type": "string",
            "description": "任务状态：待生效 | 生效中 | 已过期"
          },
          "scope": {
            "type": "array",
            "items": { "type": "string" },
            "description": "作用域列表"
          },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "description": "规则内容（根据 type 不同结构不同）"
          }
        }
      },
...
      "CountdownScheduleItem": {
        "type": "object",
        "required": ["name", "date", "priority"],
        "properties": {
          "name": { "type": "string", "description": "倒数日名称" },
          "date": { "type": "string", "format": "date", "description": "目标日期（YYYY-MM-DD）" },
          "priority": { "type": "integer", "description": "优先级，数字越大优先级越高" }
        }
      },
      "CountdownPayload": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "可选；不传时服务端根据 scope+schedules 自动生成" },
          "scope": { "$ref": "#/components/schemas/ScopeField" },
          "schedules": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CountdownScheduleItem" }
          }
        }
      },
      "CountdownRecordView": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "scope": {
            "type": "array",
            "items": { "type": "string" },
            "description": "作用域列表"
          },
          "schedules": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CountdownScheduleItem" }
          }
        }
      },
      "CountdownListResponse": {
        "type": "object",
        "required": ["loading", "hasConfig", "data"],
        "properties": {
          "loading": { "type": "boolean" },
          "hasConfig": { "type": "boolean" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CountdownRecordView" }
          }
        }
      },
...
      "ErrorWithErrorField": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "DetailResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      },
...
      "UserInfo": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "school_w", "grade_w", "class_w", "readonly"] },
          "scope": { "type": "string" }
        }
      },
      "UserInfoFull": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "school_w", "grade_w", "class_w", "readonly"] },
          "scope": { "type": "string" },
          "must_change_pwd": { "type": "boolean" },
          "must_change_username": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "AutorunPutSuccess": {
        "type": "object",
        "required": ["status", "id"],
        "properties": {
          "status": { "type": "integer", "example": 200 },
          "id": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "（已废弃，仅供旧版客户端兼容）用户名：`AstraSchedule` 或 `ElectronClassSchedule`，密码：配置文件 `secret.token`"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，通过 POST /web/auth/login 获取。写入接口还需在 X-Verify-Password 头中提供密码。"
      }
    }
  },
  "servers": [
    {
      "url": "https://{hostvar1}.getastra.cn"
    }
  ],
  "security": []
}